Using jest.mock for Usertest
